Hypertransform.jl

Welcome to the documentation for Hypertransform.jl! The package implements the two trasformations as described in [1].

Hypertriangulation

Hypertransform.hypertriangulate — Function

hypertriangulate(x::AbstractArray; bounds::Tuple{<:Real, <:Real} = (0.0, 1.0))

Maps a vector or matrix from hypercube space to hypertriangle space, scaling each element to the specified bounds.

The hypercube is the space the samplers usually work in; the components of x are in no particular order.

The hypertriangle is the space where the components are sorted into
ascenting order, ( y0 < y1 < ... < yn ). 

The (unit) transformation is defined by:

.. math::
    y_j = 1 - \prod_{i=0}^{j} (1 - x_i)^{1/(n-i)}

Example application. If we are analysing a number num_dim of DWD 
sources, all with identical priors. Then this function would be
called on the array `np.array([f_1, f_2, ..., f_num_sources])` with
`bounds=(f_min, f_max)`.

Arguments

x: An array, either a vector of length n or a matrix of size (m, n), where each row represents a point in hypercube space.

bounds: A tuple (lower, upper) specifying the minimum and maximum values for the output range. Defaults to (0.0, 1.0).

Returns

An array of the same shape as x, with each element transformed to lie within the specified bounds in hypertriangle space.

Example

using Hypertransform
x = [0.5, 0.75]
bounds = (0.0, 1.0)
y = hypertriangulate(x, bounds=bounds)

Raises

ArgumentError: If any element of x is outside the specified bounds.

Hypercubification

Hypertransform.hypercubify — Function

hypercubify(y::AbstractArray; bounds::Tuple{<:Real, <:Real} = (0.0, 1.0))

Transform a vector or matrix from the hypertriangle space back to the hypercube.

Arguments

y: A vector of length n or a matrix of size (m, n) where each row is a point in the hypertriangle space.

bounds: A tuple (lower, upper) defining the target parameter bounds. Defaults to (0.0, 1.0).

Returns

Transformed vector or matrix of the same shape as input, mapped back to the hypercube space.

References

[1]

R. Buscicchio, E. Roebber, J. M. Goldstein and C. J. Moore. Label switching problem in Bayesian analysis for gravitational wave astronomy. Phys.Rev.D 100, 084041 (2019), arXiv:1907.11631 [astro-ph.IM].